requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# NLP and Analysis
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
nltk>=3.8.1

# Data storage
sqlalchemy>=2.0.0

# API clients
newsapi-python>=0.2.7

# Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
